Map -> Level scripts, introduced separate namespace for level stuff
