--- Assets/Scripts/Map/FinishPoint.cs
@@ -1,17 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class FinishPoint : MonoBehaviour
-{
-    public string loadSceneName = "MapGenerator";
-
-    private void OnTriggerEnter(Collider other)
-    {
-        if (other.CompareTag("Player") == false)
-            return;
-
-        SceneManager.LoadScene(loadSceneName);
-    }
-}
