+++ Assets/Scripts/Utility/BareTransform.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace YAGTSS.Utility
+{
+    public struct BareTransform
+    {
+        public Vector3 position;
+        public Quaternion rotation;
+    }
+}
