+++ Assets/Scripts/Utility/BareTransform.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 336e9ee20e83c3444bcc53c79d471490
+timeCreated: 1494106886
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
