+++ Assets/Scripts/Utility/CollectionExtensions.cs
@@ -1,0 +1,30 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+
+namespace YAGTSS.Utility.Extensions
+{
+    public static class MyExtensions
+    {
+        private static System.Random rng = new System.Random();
+
+        public static void Shuffle<T>(this IList<T> list)
+        {
+            int n = list.Count;
+            while (n > 1)
+            {
+                n--;
+                int k = rng.Next(n + 1);
+                T value = list[k];
+                list[k] = list[n];
+                list[n] = value;
+            }
+        }
+
+        public static IEnumerable<T> Shuffle<T>(this IEnumerable<T> source)
+        {
+            var array = source.ToArray();
+            array.Shuffle();
+            return array;
+        }
+    }
+}
