+++ Assets/Scripts/Utility/CollectionExtensions.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: d0e70da5f0e01434c9f70f4ce93b8bc9
+timeCreated: 1494106643
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
